Share single EquatableArray<T> across source generators (#9105)
