+++ src/Analyzers/MSTest.SourceGeneration/MSTest.SourceGeneration.csproj
@@ -42,6 +42,7 @@ This package provides the C# source generators that emit reflection metadata so 
 
   <ItemGroup>
     <Compile Include="$(RepoRoot)src/Polyfills/**/*.cs" Link="Polyfills\%(RecursiveDir)%(Filename)%(Extension)" />
+    <Compile Include="..\Shared\EquatableArray.cs" Link="Shared\EquatableArray.cs" />
   </ItemGroup>
 
   <ItemGroup>
